More database logic. But there's a mutability bug.
There's a bug where looking something up from the database returns a mutable copy of that object, so if you try to adorn it, things get sloppy.

But when I tried to fix that and return deep copies of the objects from the database, my App.jsx user login Effect cycled infinitely.