Input sliders never get null values.
I had a bug where when I went from a post to the profile page to the profile edit page, the console would show an error that I was turning an uncontrolled input into a controlled one, and input values should never be null.  I guess I should initialize those things to non-null values when I create them with useState!