For now, images are links to the image.
One day, they could use some sort of "full screen image viewer", but I don't want to make one tonight, or find one.  This method is sub-optimal because it requires people to use the back button, and they can't just click on the X and stay in the app when they're done.  This sounds like a post-Wednesday problem.