Modal functionality improved.
You can close the modal by clicking outside and not lose your changes.

If you click cancel, you can cancel the editing of the avatar: It will get reset to the way it was and the modal will close.

If you click cancel without having made any changes, you won't be asked to confirm.