Spoilers can round-trip thru the editor.
But I'm not satisfied that I've found the root cause of the problem I'm having earlier.

There was some sort of "shouldJoin" visitor that was matching on the spoiler, and it visitors.find never checked the spoiler one because this shouldJoin visitor matched first.  Instead of figuring out why the shouldJoin visitor was matching the spoiler node, I just set the priority of the spoiler visitor so it would check that first.

I would like to get to the bottom of this and figure out why the shouldJoin one wanted to join them.  I think this has something to do with Lexical's TextNode, and it wants to join adjacent TextNode's together.  My SpoilerNode extends TextNode, but maybe doesn't do enough to differentiate itself.

BTW, all this Visitor stuff is MDXEditor stuff, not Lexical stuff.

Look at MDXEditor/src/exportMarkdownFromLexical.ts:158, where it says
```
const visitor = visitors.find((visitor) => visitor.testLexicalNode?.(lexicalNode))
```

That's the thing that finds the shouldJoin visitor instead of the Spoiler visitor.

Figure out what that shouldJoin visitor is doing and why it matches.  Put a breakpoint there.