Spoiler is an ElementNode, not a TextNode, in Lexical.
I was having problems before with spoilers not being tested-for because the test for whether it was a TextNode was happening first, and all that test asked was whether node was instanceof TextNode, and it was, because SpoilerNode was a subclass of TextNode.  I got around it by increasing the priority of the SpoilerNode test, but that felt hacky and I wanted to understand Lexical better.

So now, SpoilerNode subclasses ElementNode instead, and doesn't specify a priority, and now it gets detected properly.