It's actually easier if spoiler extends TextNode.
When I had it extend ElementNode, I was supposed to define all sorts of behaviors where I just wanted it to act like a text node.  Selection, adding and deleting text, etc.  I'll just make it a text node and learn that stuff when I have a more complicated type of node to create.

It might be worth looking into how Lexical's LinkNode works.  It's a text-like node that extends ElementNode.  It looks like maybe what it does is have a TextNode *inside* it??