Doing more typescript.
When I try to copy the spoiler node from the editor, it says it doesn't implement exportJSON even though it does, so I thought maybe there was a typescript problem??  So I tried to import a bunch of typescript stuff and run tsc to see what errors it gives me.  It gives me errors I do not understand, but I don't see why it would single out that one function to say wasn't implemented.  That function isn't especially mentioned in the errors I get.