Starting to versionify the posts database.
The posts database has been split into "posts" and "postVersions", and postVersions is the thing that has the text of each version.

So far, we only have features that care about the latest version of a post, so we just rewrote logic/posts.js to fetch the latest version of the post and return it in the normal format, as though that's the only version.  We will create new features that require checking the different versions of a post.

We haven't checked if creating a post or reply still works.  It probably doesn't.

Once that's working, we can create some edit functionality.

We also don't have any delete functionality.