You can delete posts.
They immediately disappear from the thread page, and they get marked as deleted in the database.

However, nothing actually takes deleted posts into account -- they are displayed just like any other.

What we should do is, if a deleted post doesn't have any replies, we just omit it from lists we return from the postsDB.  But if it does have replies, it should be displayed as a tombstone, like "(deleted post)", and don't show who wrote it and don't show the statistics.

Displaying as a tombstone should be done in Post.jsx.  Omitting them from results should be done by logic/posts.js

What also needs to be done:
I think we'll need special handling for deleting the first post of a thread, because there's no replies array to remove it from.

Also, we probably need to do something special when you delete the post that is the main post.  Maybe navigate you to the post it was a reply to?  Maybe show the post as a tombstone, even if it has no replies?