New way of filling test data.
We need this because we're filling the test data with stuff that's fetched asynchronously, which we can't do from WITHIN the onupgradeneeded transaction:  Any attempt to await, there, will cause the transaction to be closed when we come back.

So we're setting our own flag to say when the test data has been added, and we're doing that asynchronously, so we can fill the test data with stuff fetched asynchronously.  To wit, images, which we will fetch with fetch(), and then turn into blobs and ArrayBuffers and hashes.