Thread-line highlighting by javascript.
It turns out it was causing massive slowdowns to be chucking all that computed-on-the-fly stuff into the <style> tag to highlight threadlines.  So instead, we are doing the highlighting by having javascript hand out css classes, and then having simple, static rules to display the highlighting.

This implementation does not yet do minimap highlighting.