Getting closer to follower updates that make sense.
We should have a separate cursor for backfilling, so we can backfill all the followers.  Then a separate cursor for forward-filling.

When I say "cursor", I mean the person object should hold the database ids of the earliest known and latest known followers (and followees, separately).  If we detect there's backfilling to do, we backfill.  If we detect there's front-filling to do, we front-fill.