Refactor for clarity. Also, parallellism is not available.
I changed the name of chunkSource to knownChunks for clarity.

Also, I thought you could use a "skip" parameter to load the next 3 pages at a time, but it turns out there is no "skip" parameter in the mastodon api.  You have to get a page before you can know the address of the next page.